first commit
This commit only conains code, with hard-coded extensions. Tensorflow is required to run it, as well as many other files that enable the use of this code. I have chosen to commit this to a public repository to show what sort of work goes behind developing satellite data processing algorithms for the space sector.